+++ t/Venus_Run.t
@@ -2006,28 +2006,12 @@ the system with a new configuration file.
   #     info => '$PERL -V',
   #     lint => 'perlcritic',
   #     okay => '$PERL -c',
-  #     repl => '$PERL -dE0',
+  #     repl => '$REPL',
   #     reup => 'cpanm -qn Venus',
   #     says => 'eval "map log(eval), @ARGV"',
   #     shim => '$PERL -MVenus=true,false,log',
   #     test => '$PROVE',
   #     tidy => 'perltidy',
-  #   },
-  #   flow => {
-  #     deps => [
-  #       'cpan Perl::Critic',
-  #       'cpan Perl::Tidy',
-  #       'cpan Pod::Perldoc',
-  #     ],
-  #     prep => [
-  #       'deps',
-  #       'reqs',
-  #     ],
-  #     reqs => [
-  #       'which perlcritic',
-  #       'which perldoc',
-  #       'which perltidy',
-  #     ],
   #   },
   #   libs => [
   #     '-Ilib',
@@ -2045,6 +2029,7 @@ the system with a new configuration file.
   #   vars => {
   #     PERL => 'perl',
   #     PROVE => 'prove'
+  #     REPL => '$PERL -dE0'
   #   },
   # }