+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version: 3.40
+
+- [feature] Implement Venus#resolve
+- [feature] Implement automatic Venus::Error message formatting
+- [feature] Implement Venus::String#format
+- [feature] Implement Venus#{docs,text}
+- [feature] Implement Venus::{Array,Hash}#merge via Venus#merge
+- [update] Fix Venus::Run variable interpolation issue
+- [update] Resolve Venus::Run "which" issues caused by var-path-lib-var
+- [update] Resolve CPANTS issues from 3.30
+- [update] Update Venus::{Array,Hash}#{get,set} for k/v and whole-values
+
 ## Version: 3.30
 
 - [feature] Implement Venus::Container